ref:overall changes

+++ src/controllers/products/product-controller.js
@@ -1,8 +1,10 @@
+import mongoose from 'mongoose'
 import Product from '../../models/product.model'
 import * as response from '../../utils/response-util'
 
 export const createProduct = async (req, res) => {
     try {
+
         const product = await Product.create(req.body)
         return response.sendSuccess(res, 200, 'product created', [product])
     } catch (error) {
@@ -11,15 +13,34 @@ export const createProduct = async (req, res) => {
 }
 export const listProduct = async (req, res) => {
     try {
-        const productList = await Product.find({}, { 'createdAt': 0, 'updatedAt': 0, '__v': 0 })
+        //return res.send('123')
+        const { categoryId,product} = req.body
+        if(categoryId &&(!mongoose.Types.ObjectId.isValid(categoryId))){
+            return response.sendError(res,400,'send valid id')
+        }
+        let categoryFilter = {}
+        if(!categoryId){
+            categoryFilter = {}
+        }
+        else{
+            categoryFilter.categoryId=categoryId
+        }
+        //console.log(categoryFilter)
+        const productList = await Product.find(categoryFilter, { 'createdAt': 0, 'updatedAt': 0, '__v': 0 })
             .populate({
-                path: 'category', select: { '__v': 0, 'createdAt': 0, 'updatedAt': 0 }
-            })
-            .populate({
-                path: 'farmer', select: { '__v': 0, 'password': 0, 'role': 0 }
+                path: 'productId', match: {
+                    $or: [
+                        { product: { $regex: product, $options: 'i' } },
+                    ]
+                }, select: { '__v': 0, 'createdAt': 0, 'updatedAt': 0 },
+                populate: {
+                    path: 'category', select: { '__v': 0, 'createdAt': 0, 'updatedAt': 0 }
+                }
             })
 
-        return response.sendSuccess(res, 200, 'category list', productList)
+            const filterData=productList.filter(data=>data.productId!=null)
+
+        return response.sendSuccess(res, 200, 'category list', filterData)
     } catch (error) {
         return response.sendError(res, 500, error.message)
     }
@@ -27,17 +48,20 @@ export const listProduct = async (req, res) => {
 
 export const productDetails = async (req, res) => {
     try {
+        
         if (Object.keys(req.query).length == 0) {
             return response.sendError(res, 400, 'pass id to get product details')
         }
-        const productDetails = await Product.findById(req.query.productId)
+        if(!mongoose.Types.ObjectId.isValid(req.query.productId)){
+            return response.sendError(res,400,'send valid id')
+        }
+        const productDetails = await Product.findOne(req.query, { 'createdAt': 0, 'updatedAt': 0, '__v': 0 })
             .populate({
-                path: 'category', select: { '__v': 0, 'createdAt': 0, 'updatedAt': 0 }
+                path: 'productId', select: { '__v': 0, 'createdAt': 0, 'updatedAt': 0 },
+                populate: {
+                    path: 'category', select: { '__v': 0, 'createdAt': 0, 'updatedAt': 0 }
+                }
             })
-            .populate({
-                path: 'farmer', select: { '__v': 0, 'password': 0, 'role': 0 }
-            })
-            .select('-__v -createdAt -updatedAt')
         if (!productDetails) {
             return response.sendError(res, 400, 'no product found by id')
         }
@@ -63,8 +87,12 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     try {
+        return res.send('123')
         if (Object.keys(req.query).length == 0) {
             return response.sendError(res, 400, 'send id to delete')
+        }
+        if (!mongoose.Schema.Types.ObjectId.isValid(req.query.productId)) {
+            return response.sendError(res, 500, 'send valid id')
         }
         const deleteProduct = await Product.findByIdAndDelete(req.query.productId)
         if (!deleteProduct) {
